+++ src/features/seasoning/usecases/GetSeasoningListUseCase.ts
@@ -1,5 +1,5 @@
-import type { ISeasoningRepository } from "@/libs/database/interfaces/ISeasoningRepository";
-import type { ISeasoningTypeRepository } from "@/libs/database/interfaces/ISeasoningTypeRepository";
+import type { ISeasoningRepository } from "@/libs/database/interfaces";
+import type { ISeasoningTypeRepository } from "@/libs/database/interfaces";
 import { SeasoningService } from "@/features/seasoning/services/SeasoningService";
 import { sortSeasoningsByExpiry } from "@/utils/seasoningSort";
 import { calculateExpiryStatus } from "@/utils/expiryCalculation";
